Fix Hanging if defaultRepo is not defined
If defaultRepo was not defined in the config, and a user had
posted an issue with the syntax (gh#1234), then the plugin
would cause NodeBB to hang.

This fix introduces a check so that if a defaultRepo is not
defined and that syntax is used, then that match is thrown
out.

Also, the nbbpm.compatibility value was updated.